pyproject.toml
# Spide Agent — 热点新闻抓取 Agent CLI
# Copyright (C) 2026 IoTchange - All Rights Reserved
# Author: 外星动物（常智） / IoTchange / 14455975@qq.com
# Harness Engineering 架构，基于 OpenHarness 模式改造

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spide-agent"
version = "1.1.1"
description = "热点新闻信息抓取与智能整理 Agent CLI 工具"
readme = "README.md"
license = "Proprietary"
requires-python = ">=3.12"
authors = [
    { name = "外星动物（常智）", email = "14455975@qq.com" },
]
maintainers = [
    { name = "IoTchange", email = "14455975@qq.com" },
]
keywords = ["spider", "agent", "cli", "news", "hot-topics", "mcp", "mqtt"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "License :: Other/Proprietary",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    "typer>=0.12",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "rich>=13.0",
    "aiosqlite>=0.20",
    "redis[hiredis]>=5.0",
    "aiohttp>=3.9",
    "beautifulsoup4>=4.12",
    "zai-sdk>=0.2",
    "uapi-sdk-python",
    "aiomqtt>=2.0",
    "mcp>=1.0",
    "openpyxl>=3.1.5",
    "jieba>=0.42.1",
    "wordcloud>=1.9.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "aioresponses>=0.7",
]

[project.scripts]
spide = "spide.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["spide"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = [
    "RUF001", "RUF002", "RUF003",  # 中文内容中的全角标点是正常的
    "N817",                         # CamelCase import alias (WC, MockAdapter)
    "N806",                         # Mock 变量命名约定 (MockAdapter, MockClient)
    "E501",                         # 行长度，由 formatter 自动处理
]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
disallow_any_generics = false
warn_unused_configs = true
strict_optional = false
no_implicit_optional = true
warn_redundant_casts = true
warn_return_any = false
disallow_untyped_calls = false

# 第三方库缺少类型 stubs
[[tool.mypy.overrides]]
module = [
    "zai.*", "uapi.*", "aiomqtt.*", "mcp.*", "openpyxl.*",
    "wordcloud.*", "jieba.*", "yaml.*", "bs4.*",
    "aiohttp.*", "redis.*", "aiosqlite.*", "structlog.*",
]
ignore_missing_imports = true

# aiohttp session.get() kwargs 类型签名与实际用法不匹配
[[tool.mypy.overrides]]
module = ["spide.spider.fetcher"]
disable_error_code = ["arg-type"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::pytest.PytestUnraisableExceptionWarning",
]
markers = [
    "integration: 集成测试（需要网络，真实 API 调用）",
    "e2e: 端到端测试（完整 CLI 流程）",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "types-pyyaml>=6.0.12.20260408",
]
